[DGP] DPP for DGP, all tests passing.
Implements DPP for DGP. The DPP ruleset for DGP is analogous to DPP for DCP:
the program must be DGP in both the parameters and the variables, with
an exception for the power atom (which is an analog of multiplication in DCP).
Specifically, an expression `expr**parameter` is considered DPP (log-log affine)
as long as `expr` is not parametrized. For simplicity, the parameter cannot
be replaced with a log-log affine composition of parameters, ie it must be a
raw parameter.

This change requires delaying the rational approximation to the exponent of the
power atom required by Dcp2Cone. That's now done in a dcp2cone reduction
(power_canon), instead of at initialization. This leads to a slight semantic
difference in power's monotonicty calculation, which was previously based
on the rational approximation to the exponent p. With this change, the
monotonicity depends on the original value.